extract EpochIntervalTick to lib
